refactor: move human task package to top level
